Added a condition for Firefox not existing

+++ apply.sh
@@ -11,9 +11,10 @@ link .termux termux.properties
 link . .inputrc
 link . .gitconfig
 
-FIREFOX_PATH="$(python get_firefox_default_profile_path.py)"
-FIREFOX_CHROME_PROFILE_FILE_NAME='userChrome.css'
-FIREFOX_CHROME_PATH="$FIREFOX_PATH/chrome"
-rm "$FIREFOX_CHROME_PATH"/"$FIREFOX_CHROME_PROFILE_FILE_NAME"
-mkdir -p "$FIREFOX_CHROME_PATH"
-ln -rs "firefox/$FIREFOX_CHROME_PROFILE_FILE_NAME" "$FIREFOX_CHROME_PATH/$FIREFOX_CHROME_PROFILE_FILE_NAME"
+if FIREFOX_PATH="$(python get_firefox_default_profile_path.py)" ; then
+    FIREFOX_CHROME_PROFILE_FILE_NAME='userChrome.css'
+    FIREFOX_CHROME_PATH="$FIREFOX_PATH/chrome"
+    rm "$FIREFOX_CHROME_PATH"/"$FIREFOX_CHROME_PROFILE_FILE_NAME"
+    mkdir -p "$FIREFOX_CHROME_PATH"
+    ln -rs "firefox/$FIREFOX_CHROME_PROFILE_FILE_NAME" "$FIREFOX_CHROME_PATH/$FIREFOX_CHROME_PROFILE_FILE_NAME"
+fi
